backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
itsdangerous>=2.1.2
Pillow>=9.0.0
numpy>=1.20.0
rasterio>=1.3.0
